fixed broken test in py2-win, moved some tests
